// frontend/src/app/model/period.ts
import { Course } from './course';

export class Period {
  periodId: number;
  section: string;
  semester: number;
  department: string;
  dayOfWeek: string;
  startTime: string;
  c_id: number;
  course_Id: string;
  teachersId: number;
  t_name: string[];
  durationMinute: number;
  roomId: number;
  span: number;
}